+++ src/routes/auth/change-email/request-verification.ts
@@ -7,6 +7,7 @@ import { asyncWrapper } from '@shared/helpers'
 import { EMAILS_ENABLE, SERVER_URL } from '@shared/config'
 import { emailClient } from '@shared/email'
 import { request } from '@shared/request'
+import { SetNewEmailData } from '@shared/types'
 
 import { getRequestInfo } from './utils'
 
@@ -36,8 +37,10 @@ async function requestChangeEmail(req: Request, res: Response): Promise<unknown>
     return Boom.badImplementation('Unable to set new ticket')
   }
   // set new email
+  let display_name
   try {
-    await request(setNewEmail, { user_id, new_email })
+    const setNewEmailReturn = await request<SetNewEmailData>(setNewEmail, { user_id, new_email })
+    display_name = setNewEmailReturn.update_auth_accounts.returning[0].user.display_name
   } catch (error) {
     console.error(error)
     throw Boom.badImplementation('unable to set new email')
@@ -46,7 +49,11 @@ async function requestChangeEmail(req: Request, res: Response): Promise<unknown>
   try {
     await emailClient.send({
       template: 'change-email',
-      locals: { ticket, url: SERVER_URL },
+      locals: {
+        ticket,
+        url: SERVER_URL,
+        display_name
+      },
       message: {
         to: new_email,
         headers: {