+++ src/shared/queries.ts
@@ -11,6 +11,7 @@ const accountFragment = gql`
     }
     user {
       id
+      display_name
       ${JWT_CUSTOM_FIELDS.join('\n\t\t\t')}
     }
     is_anonymous
@@ -245,6 +246,11 @@ export const setNewEmail = gql`
       where: { user: { id: { _eq: $user_id } } }
       _set: { new_email: $new_email }
     ) {
+      returning {
+        user {
+          display_name
+        }
+      }
       affected_rows
     }
   }