+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Next release
 
+- server: `display_name` is attached for email templates
+
 ## v2.1.1
 
 - server: Support streaming content (#307)